+++ src/main/java/io/github/torand/openapi2java/collectors/ResourceInfoCollector.java
@@ -16,7 +16,11 @@
 package io.github.torand.openapi2java.collectors;
 
 import io.github.torand.openapi2java.generators.Options;
-import io.github.torand.openapi2java.model.*;
+import io.github.torand.openapi2java.model.AnnotationInfo;
+import io.github.torand.openapi2java.model.ConstantValue;
+import io.github.torand.openapi2java.model.MethodInfo;
+import io.github.torand.openapi2java.model.ResourceInfo;
+import io.github.torand.openapi2java.model.SecurityRequirementInfo;
 import io.swagger.v3.oas.models.Operation;
 import io.swagger.v3.oas.models.PathItem;
 import io.swagger.v3.oas.models.security.SecurityRequirement;
@@ -31,8 +35,12 @@ import static io.github.torand.javacommons.collection.CollectionHelper.isEmpty;
 import static io.github.torand.javacommons.collection.CollectionHelper.nonEmpty;
 import static io.github.torand.javacommons.lang.StringHelper.isBlank;
 import static io.github.torand.javacommons.lang.StringHelper.nonBlank;
-import static io.github.torand.openapi2java.collectors.Extensions.*;
-import static io.github.torand.openapi2java.utils.StringUtils.getClassNameFromFqn;
+import static io.github.torand.openapi2java.collectors.Extensions.EXT_RESTCLIENT_CONFIGKEY;
+import static io.github.torand.openapi2java.collectors.Extensions.EXT_RESTCLIENT_HEADERS;
+import static io.github.torand.openapi2java.collectors.Extensions.EXT_RESTCLIENT_HEADERSFACTORY;
+import static io.github.torand.openapi2java.collectors.Extensions.EXT_RESTCLIENT_PROVIDERS;
+import static io.github.torand.openapi2java.collectors.Extensions.extensions;
+import static io.github.torand.openapi2java.utils.PackageUtils.getClassNameFromFqn;
 import static java.util.Collections.emptyList;
 import static java.util.Objects.isNull;
 import static java.util.Objects.nonNull;