+++ src/main/java/io/github/torand/openapi2java/collectors/TypeInfoCollector.java
@@ -25,8 +25,11 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import static io.github.torand.javacommons.collection.CollectionHelper.isEmpty;
+import static io.github.torand.javacommons.collection.CollectionHelper.nonEmpty;
 import static io.github.torand.javacommons.lang.Exceptions.illegalStateException;
 import static io.github.torand.javacommons.lang.StringHelper.nonBlank;
+import static io.github.torand.javacommons.stream.StreamHelper.streamSafely;
 import static io.github.torand.openapi2java.collectors.Extensions.EXT_JSON_DESERIALIZER;
 import static io.github.torand.openapi2java.collectors.Extensions.EXT_JSON_FORMAT;
 import static io.github.torand.openapi2java.collectors.Extensions.EXT_JSON_SERIALIZER;
@@ -35,7 +38,7 @@ import static io.github.torand.openapi2java.collectors.Extensions.EXT_VALIDATION
 import static io.github.torand.openapi2java.collectors.Extensions.extensions;
 import static io.github.torand.openapi2java.collectors.TypeInfoCollector.NullabilityResolution.FORCE_NOT_NULLABLE;
 import static io.github.torand.openapi2java.collectors.TypeInfoCollector.NullabilityResolution.FORCE_NULLABLE;
-import static io.github.torand.openapi2java.utils.StringUtils.getClassNameFromFqn;
+import static io.github.torand.openapi2java.utils.PackageUtils.getClassNameFromFqn;
 import static io.github.torand.openapi2java.utils.StringUtils.joinCsv;
 import static java.lang.Boolean.TRUE;
 import static java.util.Objects.nonNull;