+++ src/main/java/io/github/torand/openapi2java/writers/kotlin/KotlinPojoWriter.java
@@ -27,11 +27,13 @@ import java.io.Writer;
 import java.util.List;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Predicate;
 
 import static io.github.torand.javacommons.collection.CollectionHelper.isEmpty;
 import static io.github.torand.javacommons.collection.CollectionHelper.nonEmpty;
 import static io.github.torand.javacommons.stream.StreamHelper.streamSafely;
 import static io.github.torand.openapi2java.utils.KotlinTypeMapper.toKotlinNative;
+import static io.github.torand.openapi2java.utils.PackageUtils.isFqnInPackage;
 import static java.util.function.Predicate.not;
 
 /**
@@ -83,8 +85,10 @@ public class KotlinPojoWriter extends BaseWriter implements PojoWriter {
     }
 
     private void writeImports(PojoInfo pojoInfo) {
+        Predicate<String> isInSamePackage = fqn -> isFqnInPackage(fqn, opts.getModelPackage(pojoInfo.modelSubpackage()));
+
         List<String> imports = pojoInfo.aggregatedNormalImports().stream()
-            .filter(ni -> !isInPackage(ni, pojoInfo.modelSubpackage()))
+            .filter(not(isInSamePackage))
             .filter(not("java.util.List"::equals))
             .filter(not("java.util.Map"::equals))
             .map("import %s"::formatted)
@@ -135,14 +139,6 @@ public class KotlinPojoWriter extends BaseWriter implements PojoWriter {
         }
 
         return "@field:"+annotation.substring(1);
-    }
-
-    private boolean isInPackage(String qualifiedType, String pojoModelSubpackage) {
-        // Remove class name from qualifiedType value
-        int lastDotIdx = qualifiedType.lastIndexOf(".");
-        String typePackage = qualifiedType.substring(0, lastDotIdx);
-
-        return opts.getModelPackage(pojoModelSubpackage).equals(typePackage);
     }
 
     private static String escapeReservedKeywords(String name) {